결제 구현중

+++ src/test/java/kr/hhplus/be/server/ServerApplicationTests.java
@@ -1,11 +1,11 @@
 package kr.hhplus.be.server;
 
-import kr.hhplus.be.server.domain.PointEntity;
-import kr.hhplus.be.server.domain.PointHistoryEntity;
-import kr.hhplus.be.server.application.PointService;
-import kr.hhplus.be.server.infrastructure.PointRepository;
-import kr.hhplus.be.server.infrastructure.PointHistoryRepository;
-import kr.hhplus.be.server.domain.TransactionType;
+import kr.hhplus.be.server.domain.point.PointEntity;
+import kr.hhplus.be.server.domain.point.PointHistoryEntity;
+import kr.hhplus.be.server.application.point.PointService;
+import kr.hhplus.be.server.infrastructure.point.PointRepository;
+import kr.hhplus.be.server.infrastructure.point.PointHistoryRepository;
+import kr.hhplus.be.server.domain.point.TransactionType;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -13,6 +13,7 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.Date;
 import java.util.List;
@@ -35,13 +36,13 @@ class ServerApplicationTests {
 		String userId = "abc";
 		PointEntity fakePoint = new PointEntity();
 		fakePoint.setUserId(userId);
-		fakePoint.setPointTotal(1000L);
+		fakePoint.setPointTotal(BigDecimal.valueOf(1000L));
 
 		// 유저아이디로 포인트를 조회할때
 		when(pointRepository.findByUserId(userId)).thenReturn(fakePoint);
 
 		// 결과는 1000이 조회돼야 한다.
-		long result = pointService.getUserPoint(userId);
+		BigDecimal result = pointService.getUserPoint(userId);
 
 		// 조회되는값이 1000과 같은가?
 		assertEquals(1000L, result);
@@ -55,13 +56,13 @@ class ServerApplicationTests {
 		PointHistoryEntity history1 = new PointHistoryEntity();
 		history1.setUserId(userId);
 		history1.setType(TransactionType.CHARGE);
-		history1.setAmount(1000);
+		history1.setAmount(BigDecimal.valueOf(1000));
 		history1.setCurrentTime(new Date());
 
 		PointHistoryEntity history2 = new PointHistoryEntity();
 		history2.setUserId(userId);
 		history2.setType(TransactionType.USE);
-		history2.setAmount(500);
+		history2.setAmount(BigDecimal.valueOf(500));
 		history2.setCurrentTime(new Date());
 
 		List<PointHistoryEntity> fakeHistoryList = List.of(history1, history2);
@@ -85,7 +86,7 @@ class ServerApplicationTests {
 
 		PointEntity fakePoint = new PointEntity();
 		fakePoint.setUserId(userId);
-		fakePoint.setPointTotal(1000L);
+		fakePoint.setPointTotal(BigDecimal.valueOf(1000L));
 
 		// 현재 포인트 조회 mock
 		when(pointRepository.findByUserId(userId)).thenReturn(fakePoint);
@@ -93,7 +94,7 @@ class ServerApplicationTests {
 		// when & then: 예외가 발생하는지 확인
 		IllegalArgumentException exception = assertThrows(
 				IllegalArgumentException.class,
-				() -> pointService.chargeUserPoint(userId, amount)
+				() -> pointService.chargeUserPoint(userId, BigDecimal.valueOf(amount))
 		);
 
 		// 예외 메시지 확인
@@ -108,7 +109,7 @@ class ServerApplicationTests {
 		// 현재 포인트 만들기
 		PointEntity fakePoint = new PointEntity();
 		fakePoint.setUserId(userId);
-		fakePoint.setPointTotal(1000L);
+		fakePoint.setPointTotal(BigDecimal.valueOf(1000L));
 
 		// 포인트를 더해서 업데이트
 		long newPoint = 1000L + amount;
@@ -116,12 +117,12 @@ class ServerApplicationTests {
 		// 포인트 업데이트
 		PointEntity newFakePoint = new PointEntity();
 		newFakePoint.setUserId(userId);
-		newFakePoint.setPointTotal(newPoint);
+		newFakePoint.setPointTotal(BigDecimal.valueOf(newPoint));
 
 		// 충전된 내역을 히스토리에 인서트
 		PointHistoryEntity fakeHistory = new PointHistoryEntity();
 		fakeHistory.setUserId(userId);
-		fakeHistory.setAmount(newPoint);
+		fakeHistory.setAmount(BigDecimal.valueOf(newPoint));
 		fakeHistory.setType(TransactionType.CHARGE);
 		fakeHistory.setCurrentTime(new Date());
 
@@ -130,7 +131,7 @@ class ServerApplicationTests {
 		doReturn(newFakePoint).when(pointRepository).insertAndUpdate(userId, newPoint);
 
 		// 결과는 새로 업데이트된 포인트가 조회돼야 한다.
-		PointEntity result = pointService.chargeUserPoint(userId, amount);
+		PointEntity result = pointService.chargeUserPoint(userId, BigDecimal.valueOf(amount));
 		assertEquals(newPoint, result.getPointTotal());
 
 		// 포인트 히스토리 인서트
